Stage logic & bullets -Added basic stage logic -Added all player bullet types

+++ SDL Asteroids C/src/geometry.c
@@ -1,10 +1,11 @@
-#include "common.h"
+#include <stdlib.h>
+#include <math.h>
 
 #include "geometry.h"
 
 //normalizes a vector
 void normalize(Vector2* vector) {
-	float magnitude = sqrt(vector->x*vector->x + vector->y*vector->y);
+	float magnitude = sqrtf(vector->x*vector->x + vector->y*vector->y);
 	if (magnitude != 0) {
 		vector->x /= magnitude;
 		vector->y /= magnitude;
